+++ src/__test__/sanity.test.js
@@ -1,10 +1,10 @@
 import { render, within } from "@testing-library/react";
 import userEvent from "@testing-library/user-event";
 import "@testing-library/jest-dom";
-
-import Main from "./Components/Main";
 import { act } from "react-dom/test-utils";
 import "jest-canvas-mock";
+
+import Main from "../Components/Main";
 
 // Need to mock window.URL.createObjectURL
 // https://stackoverflow.com/questions/52968969/jest-url-createobjecturl-is-not-a-function
